test: move NamespaceManagement managed-by chain coverage to sequential suite
